Fixed a few errors.

+++ src/include/primer/p0_starter.h
@@ -12,6 +12,7 @@
 
 #pragma once
 
+#include <algorithm>
 #include <memory>
 #include <stdexcept>
 #include <vector>
@@ -179,7 +180,7 @@ class RowMatrix : public Matrix<T> {
    * @throws OUT_OF_RANGE if `source` is incorrect size
    */
   void FillFrom(const std::vector<T> &source) override {
-    if (source.size() != static_cast<size_t>(this->rows_ * this->cols_)) {
+    if (static_cast<int>(source.size()) != this->rows_ * this->cols_) {
       throw Exception{ExceptionType::OUT_OF_RANGE, "RowMatrix::FillFrom() out of range!"};
     }
     std::copy(source.begin(), source.end(), this->linear_);
